Add reCAPTCHA Javascript to Form Template
* Created a new handlebars template with just the required javascript
* Included that javascript template in the main form template
Add Code to Handle reCAPTCHA Submission

First pass on verifying the form submission
Change reCAPTCHA Action to 'rfi'

Part of debugging the reCAPTCHA code (the action was saying 'homepage' no matter what I did). Turns out it was the fault of Contact Form 7.
Update Plugin to Post the WordPress Way

Changed the form so that it can submit to admin-post.php, and then registered callbacks for handling the $_POST data. Also added some debugging output.
Refactor Plugin Code

Multiple Changes to the plugin code:

* Added new constants to keep track of reCAPTCHA information needed throughout the class
* Add new callbacks for handling POST data via wp_admin_post()
* Added new form endpoints for testing submissions
* Added a space after the colon in our generic error message, to make it easier to read
* Broke the POST handling functionality out into multiple methods: one for the main logic, and other methods for supporting roles (doing the reCAPTCHa stuff, actually submitting the form, etc.)
* Made us of  the WP_Error class whenever possible, as a way to handle things that might go wrong.
Add Wordpress Action to Form

In order for Wordpress to know which callback method to call when a form is submitted, you need to provide a value in an 'action' field in your form. The value ends up as part of the name of your callback method.
Set Endpoint Based on Shortcode Attribute

We already had a shortcode attribute called 'test_mode', and we used it to put a string in a hidden field ('Test' or 'Prod'). I expanded its use to also determine which endpoint we are using.

Putting `test_mode="test"` as a shortcode attribute will now select the test endpoint as well.
Add Class Var to Track Current Endpoint

Added $this->currentEndPoint to the class for setting the API endpoint to use.
Use $this-currentEndPoint When Submitting

Modified the form submission code to use our selected endpoint.
Code Cleanup

Removed some testing stuff, added a comment.
Update Form Template for Custom Thank You

Added a new hidden field, 'thank_you', that holds the path to a thank you page. The 'formUrl' value is now used exclusively to redirect back to the page after an error.
Add Code to Handle Custom Thank You Page

Updates to allow for a new short-code attribute, 'thank_you' that can hold a URL for displaying a thank you page after a successful form submission.
Store Secret Key in Database

Github politely reminded me that I should not commit secret keys to our repo, so I changed the way we are handling it:

* updated the admin form to have a field for entering the secret key, so we can store it in the database and not in our code.
* updated the code to read the secret key from the options table in the database
* Increased the timeout value on our posting to the endpoint. I was routinely finding that it took about 5 seconds - which is the default timeout value for Wordpress's wp_remote_post() method, which led to several failed submits to to timing out.