Fixes to Make Tests Happy
We had several failures/exceptions due to optional shortcode attributes being used when setting up the view. For example, if you did not provide a 'major_code_picker' attribute in the shortcode, we were still trying to put that value in the view.

I simply made blank defaults for two optional shortcode attributed, which stoped failures on tests that did not need those attributes.