Use Correct Call to get Error Message
If we forget to set up our Source ID in the plugin settings, we would end up with a blank, white page. That was because I accidentally used get_error_message() while building the error object. You can't do that, as get_error_message() is for retrieving the error message from an error object!

I Replaced that call with the correct HTTP API method, and an error message now displays, instead of a blank page.