Add Storage of Site Key to Admin Screen
To allow for different sets of keys for development and production sites (if we want to go that way), updated the admin screen to allow for storing the public Site Key from Google reCAPTCHA.

This fixes an issue with having that site key hard-coded in a Javascript file and testing across different domain names. Now you can create different keys for dev/prod domains if you  want.