Fix Logic Error with Endpoint
My initial logic for setting the endpoint was happening when we were building the form, and would NOT have affected where we actually submitted anything.

Moved, and updated, the logic to check the value of the actual $_POST variable we set, and determine endpoint based on that just before submitting the form.