Decouple Endpoint and TestMode
Turns out you can send 'Test' submissions to the production endpoint. My idea to use the 'test_mode' short code attribute to determine the endpoint was not a good idea. Allowing for an 'endpoint' attribute (defaulting to 'Prod') for specific cases where you want your form to submit to the test endpoint.

TLDR: set the short code attribute 'endpoint' to the value 'test' in order to use the test endpoint. If it's set to anything else, or not set at all, RFI submissions will go to the production endpoint.