Implement Token-Check
To prevent ContactForm7 from creating a token on page load, then having that token timeout before our users even fill out a form, I'm adding a token-check field that we will explicitly set to 'check' when we ask for our own token.

Since ContactForm7 will not have set the token-check value, we can presume that any tokens that exist without our check value set might be stale; in that case we get a new one (and set the check field to 'check').