Fix Boolean Logic Bug
We had a classic PHP bug here: 

* Accept the short-code attribute "false", which is a string
* Evaluate that string to see if we should create our Program of Interest list 
* PHP happily evaluates a non-empty string as TRUE
* We get a drop-down list 100% of the time

There is only one value (right now) that can fall prey to this issue, so I fixed it with a call to PHP's built-in filter_var() method and asked for the appropriate boolean value. Then, later, when we are deciding whether or not to compile our list, I explicitly require that 'major_code_picker' be a boolean TRUE.

This fixed the issue on my dev machine, preventing the drop-down list from being created, and still preserving the hidden POI field and it's value of "SUEMSLEMSL".

I'm not super familiar with the business logic here, and any ramifications of this change in that regard, but there clearly was a logic bug that should now be fixed.